FAC-249:feature:Move mother classes to test.fixture
