+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/application/service/advice/PlanConstraintProviderMultipleTargetTest.java
@@ -4,7 +4,7 @@ import ai.timefold.solver.test.api.score.stream.ConstraintVerifier;
 import org.flickit.assessment.advice.application.domain.AttributeLevelScore;
 import org.flickit.assessment.advice.application.domain.Plan;
 import org.flickit.assessment.advice.application.domain.Question;
-import org.flickit.assessment.advice.application.service.QuestionMother;
+import org.flickit.assessment.advice.test.fixture.application.QuestionMother;
 import org.junit.jupiter.api.Test;
 
 class PlanConstraintProviderMultipleTargetTest {
