+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/test/fixture/application/QuestionMother.java
@@ -1,9 +1,9 @@
-package org.flickit.assessment.advice.application.service;
+package org.flickit.assessment.advice.test.fixture.application;
 
 import org.flickit.assessment.advice.application.domain.AttributeLevelScore;
 import org.flickit.assessment.advice.application.domain.Question;
 
-import static org.flickit.assessment.advice.application.service.OptionMother.createOptions;
+import static org.flickit.assessment.advice.test.fixture.application.OptionMother.createOptions;
 
 public class QuestionMother {
